Solve F24 #2: Magic Beehive
After determining that the problem is too big to solve by hand, I
started working on a bruteforce solution. I tried a lot of things
(making heavy use of itertools combinations/permutations), but they were
all far too inefficient to run in a reasonable timeframe.

Ultimately, I landed on an piece-by-piece approach that starts by
recursively calculating valid outer rings (of which there are 30504) and
then moves inward. This approach doesn't make any use of itertools, so
all the math is done in Python, but enough branches are pruned that it
finishes in under a second (finding all 12 rotations/reflections of the
single unique solution).